part2/2.11/backend/requirements.txt
django
tzdata
psycopg2-binary
djangorestframework
django-cors-headers
django-filter
djangorestframework-gis
python-dateutil